browser/components/feeds/test/test_bug494328.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=494328
-->
<head>
  <title>Test for bug 494328</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=494328">Mozilla Bug 494328</a>
<p id="display"><iframe id="testFrame" src="bug494328-data.xml"></iframe></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 494328 **/
SimpleTest.waitForExplicitFinish();

addLoadEvent(function() {
  var links = SpecialPowers.wrap($("testFrame")).contentDocument.getElementById("feedContent").querySelectorAll("div.enclosure > a");
  is(links[0].textContent, "Episode 1", "filename decoded incorrectly");
  is(links[1].textContent, "Episode #2", "filename decoded incorrectly");
  is(links[2].textContent, "http://www.example.com/podcasts/Episode #3/", "filename decoded incorrectly");
  is(links[3].textContent, "Is This Episode #4?", "filename decoded incorrectly");
});
addLoadEvent(SimpleTest.finish);

</script>
</pre>
</body>
</html>
